* merge various code versions to one, driven by config
