--- Car_WS_ENG/Car_WS_ENG.ino
@@ -1,158 +1,0 @@
-// Car on ESP32-CAM 
-
-const char* ssid = "**********";
-const char* password = "**********";
-
-#include "esp_wifi.h"
-#include "esp_camera.h"
-#include <WiFi.h>
-#include "soc/soc.h"
-#include "soc/rtc_cntl_reg.h"
-
-#define CAMERA_MODEL_AI_THINKER
-
-#define PWDN_GPIO_NUM     32
-#define RESET_GPIO_NUM    -1
-#define XCLK_GPIO_NUM      0
-#define SIOD_GPIO_NUM     26
-#define SIOC_GPIO_NUM     27
-#define Y9_GPIO_NUM       35
-#define Y8_GPIO_NUM       34
-#define Y7_GPIO_NUM       39
-#define Y6_GPIO_NUM       36
-#define Y5_GPIO_NUM       21
-#define Y4_GPIO_NUM       19
-#define Y3_GPIO_NUM       18
-#define Y2_GPIO_NUM        5
-#define VSYNC_GPIO_NUM    25
-#define HREF_GPIO_NUM     23
-#define PCLK_GPIO_NUM     22
-
-void startCameraServer();
-
-const int MotPin0 = 12;  // Right motor
-const int MotPin1 = 13;  // Right motor
-const int MotPin2 = 14;  // Left motor
-const int MotPin3 = 15;  // Left motor
-
-void initMotors() 
-{
-  ledcSetup(3, 2000, 8); // 2000 hz PWM, 8-bit resolution
-  ledcSetup(4, 2000, 8); // 2000 hz PWM, 8-bit resolution
-  ledcSetup(5, 2000, 8); // 2000 hz PWM, 8-bit resolution
-  ledcSetup(6, 2000, 8); // 2000 hz PWM, 8-bit resolution
-  ledcAttachPin(MotPin0, 3); // Right motor
-  ledcAttachPin(MotPin1, 4); // Right motor
-  ledcAttachPin(MotPin2, 5); // Left motor
-  ledcAttachPin(MotPin3, 6); // Left motor
-}
-
-const int ServoPin = 2;  
-void initServo() 
-{
-  ledcSetup(8, 50, 16); // 50 hz PWM, 16-bit resolution, range from 3250 to 6500.
-  ledcAttachPin(ServoPin, 8); 
-}
-
-void setup() 
-{
-  WRITE_PERI_REG(RTC_CNTL_BROWN_OUT_REG, 0);
-  
-  Serial.begin(115200);
-  Serial.setDebugOutput(true);
-  Serial.println();
-
-  camera_config_t config;
-  config.ledc_channel = LEDC_CHANNEL_0;
-  config.ledc_timer = LEDC_TIMER_0;
-  config.pin_d0 = Y2_GPIO_NUM;
-  config.pin_d1 = Y3_GPIO_NUM;
-  config.pin_d2 = Y4_GPIO_NUM;
-  config.pin_d3 = Y5_GPIO_NUM;
-  config.pin_d4 = Y6_GPIO_NUM;
-  config.pin_d5 = Y7_GPIO_NUM;
-  config.pin_d6 = Y8_GPIO_NUM;
-  config.pin_d7 = Y9_GPIO_NUM;
-  config.pin_xclk = XCLK_GPIO_NUM;
-  config.pin_pclk = PCLK_GPIO_NUM;
-  config.pin_vsync = VSYNC_GPIO_NUM;
-  config.pin_href = HREF_GPIO_NUM;
-  config.pin_sscb_sda = SIOD_GPIO_NUM;
-  config.pin_sscb_scl = SIOC_GPIO_NUM;
-  config.pin_pwdn = PWDN_GPIO_NUM;
-  config.pin_reset = RESET_GPIO_NUM;
-  config.xclk_freq_hz = 20000000;
-  config.pixel_format = PIXFORMAT_JPEG;
-  if(psramFound()){
-    config.frame_size = FRAMESIZE_QVGA;
-    config.jpeg_quality = 10;
-    config.fb_count = 2;
-  } else {
-    config.frame_size = FRAMESIZE_QVGA;
-    config.jpeg_quality = 12;
-    config.fb_count = 1;
-  }
-
-  esp_err_t err = esp_camera_init(&config);
-  if (err != ESP_OK) {
-    Serial.printf("Camera init failed with error 0x%x", err);
-    return;
-  }
-
-  sensor_t * s = esp_camera_sensor_get();
-  s->set_framesize(s, FRAMESIZE_QVGA);
-  s->set_vflip(s, 1);
-  s->set_hmirror(s, 1);
-
-  initMotors();
-  initServo();
-  
-  ledcSetup(7, 5000, 8);
-  ledcAttachPin(4, 7);  //pin4 is LED
-  
-  Serial.println("ssid: " + (String)ssid);
-  Serial.println("password: " + (String)password);
-  
-  WiFi.begin(ssid, password);
-  delay(500);
-
-  long int StartTime=millis();
-  while (WiFi.status() != WL_CONNECTED) 
-  {
-      delay(500);
-      if ((StartTime+10000) < millis()) break;
-  }
-  
-  startCameraServer();
-
-  if (WiFi.status() == WL_CONNECTED) 
-  {
-    Serial.println("");
-    Serial.println("WiFi connected");    
-    Serial.print("Camera Ready! Use 'http://");
-    Serial.print(WiFi.localIP());
-    Serial.println("' to connect");
-  } else {
-    Serial.println("");
-    Serial.println("WiFi disconnected");      
-    Serial.print("Camera Ready! Use 'http://");
-    Serial.print(WiFi.softAPIP());
-    Serial.println("' to connect");
-    char* apssid = "ESP32-CAM";
-    char* appassword = "12345678";         
-    WiFi.softAP((WiFi.softAPIP().toString()+"_"+(String)apssid).c_str(), appassword);    
-  }
-
-  for (int i=0;i<5;i++) 
-  {
-    ledcWrite(7,10);  // flash led
-    delay(50);
-    ledcWrite(7,0);
-    delay(50);    
-  }
-}
-
-void loop() {
-  delay(1000);
-  Serial.printf("RSSi: %ld dBm\n",WiFi.RSSI()); 
-}
